Update build env: VS 18, MSVC v145, SPDX/ABI metadata
Migrated build to VS 18/Insiders and MSVC v145 toolset, updating all build scripts, paths, and toolchain references from "wor" to "worr-kex". Regenerated all CMake, Ninja, and vcpkg files to reflect new environment, including updated ABI hashes and detailed SPDX SBOM metadata for fmt, jsoncpp, vcpkg-cmake, and vcpkg-cmake-config. Improves build reproducibility and traceability.

+++ src/server/gameplay/g_clients.cpp
@@ -68,7 +68,7 @@ void AllocateClientArray(int maxClients) {
 	// [KEX]: Ensure client pointers are linked immediately to prevent engine crashes
 	// if SV_CalcPings runs before a client is fully connected.
 	if (g_entities) {
-		for (int i = 0; i < game.maxClients; i++) {
+		for (int i = 0; i < static_cast<int>(game.maxClients); i++) {
 			g_entities[i + 1].client = &game.clients[i];
 		}
 	}
@@ -77,7 +77,7 @@ void AllocateClientArray(int maxClients) {
 void FreeClientArray() {
 	// [KEX]: Unlink client pointers
 	if (g_entities && game.clients) {
-		for (int i = 0; i < game.maxClients; i++) {
+		for (int i = 0; i < static_cast<int>(game.maxClients); i++) {
 			g_entities[i + 1].client = nullptr;
 		}
 	}
